+++ src/Tile.h
@@ -8,7 +8,7 @@ public:
 	enum Type { Walkable, Solid, };
 	Type type;
 
-	Tile(const Texture& texture, const int& tileType, const glm::vec2& sourceFramePosition = glm::vec2(0.0f), const glm::vec3& position = glm::vec3(0.0f));
+	Tile(const Texture& texture, const int& tileType, const glm::ivec2& sourceFramePosition = glm::ivec2(0.0f), const glm::vec3& position = glm::vec3(0.0f));
 	~Tile(void);
 
 	void Update(const float& deltaTime);
